gate_monitor/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.9.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
